fine_tune_tweet_classifier/src/services/config_service.py
from src.utils.singleton import Singleton

import pathlib
import uuid
import yaml
import os

class ConfigService(metaclass=Singleton):
    def __init__(self, configs: pathlib.Path):
        self.config = {} 

        # Append the contents of every .yaml file in configs directory into self.config dictionary 
        for config_file in os.listdir(configs):
            config_name = config_file[:-5]  # Remove .yaml suffix
            config_path = configs.joinpath(config_file) 
            
            with open(config_path, "r") as cf:
                self.config[config_name] = yaml.safe_load(cf)

    @property
    def tweets_csv_path(self):
        return self.config["config"]["tweets_csv_path"]